+++ src/lib/dropdowns/DropdownItem.svelte
@@ -3,6 +3,7 @@
 	export let liClass: string =
 		'block font-medium cursor-pointer py-2 px-4 text-sm hover:bg-gray-100 dark:hover:bg-gray-600';
 	export let color: string = 'default';
+	export let tabindex: number = 0;
 
 	const colors = {
 		default: 'text-gray-700 dark:text-gray-200 dark:hover:text-white',
@@ -21,7 +22,6 @@
 	on:blur
 	on:mouseenter
 	on:mouseleave
-	tabindex="0"
->
+	{tabindex}>
 	<slot />
 </li>
